requirements.txt
toml >= 0.10.2
python-telegram-bot >= 20.8
